+++ src/packages/piral-core/src/hooks/search.ts
@@ -1,9 +1,9 @@
-import { useEffect, useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { wrapElement } from 'react-arbiter';
 import { useDebounce } from './debounce';
 import { useGlobalState } from './globalState';
 import { useActions } from './actions';
-import { Disposable } from '../types';
+import { Disposable, SearchOptions } from '../types';
 
 /**
  * Hook that yields the possibility of searching in Piral.
@@ -14,25 +14,37 @@ import { Disposable } from '../types';
  * All registered search providers are used and search results
  * will be integrated as they arrive.
  */
-export function useSearch(): [string, (value: string) => void] {
+export function useSearch(): [string, (value: string) => void, () => void] {
   const { setSearchInput, resetSearchResults, appendSearchResults } = useActions();
+  const [immediate, setImmediate] = useState(false);
   const searchInput = useGlobalState(m => m.search.input);
   const providers = useGlobalState(m => m.components.searchProviders);
   const q = useDebounce(searchInput);
   const cancel = useRef<Disposable>(undefined);
+  const triggerSearch = () => setImmediate(true);
+  const query = immediate ? searchInput : q;
 
   useEffect(() => {
     const providerKeys = Object.keys(providers);
     const load = !!q && providerKeys.length > 0;
+    const enter = immediate;
     cancel.current && cancel.current();
     resetSearchResults(load);
+
+    if (enter) {
+      setImmediate(false);
+    }
 
     if (load) {
       let searchCount = providerKeys.length;
       let active = true;
       cancel.current = () => (active = false);
+      const opts: SearchOptions = {
+        query,
+        immediate: enter,
+      };
       providerKeys.forEach(key =>
-        providers[key].search(q).then(
+        providers[key].search(opts).then(
           results => {
             active && appendSearchResults(results.map(m => wrapElement(m)), --searchCount === 0);
           },
@@ -43,7 +55,7 @@ export function useSearch(): [string, (value: string) => void] {
         ),
       );
     }
-  }, [q]);
+  }, [query]);
 
-  return [searchInput, setSearchInput];
+  return [searchInput, setSearchInput, triggerSearch];
 }