+++ src/packages/piral/src/types/api.ts
@@ -4,6 +4,7 @@ import {
   TileComponentProps,
   MenuComponentProps,
   ExtensionComponentProps,
+  ModalComponentProps,
 } from 'piral-core';
 import { PiralFetchApi, PiralGqlApi, PiralLocaleApi } from 'piral-ext';
 
@@ -11,5 +12,6 @@ export type PiExtApi = PiralFetchApi & PiralGqlApi & PiralLocaleApi;
 export type PiletApi = PiralApi<PiExtApi>;
 export type PiPageComponentProps = PageComponentProps<PiletApi>;
 export type PiTileComponentProps = TileComponentProps<PiletApi>;
+export type PiModalComponentProps<TOpts> = ModalComponentProps<PiletApi, TOpts>;
 export type PiMenuComponentProps = MenuComponentProps<PiletApi>;
 export type PiExtensionComponentProps<T = Record<string, any>> = ExtensionComponentProps<PiletApi, T>;