+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.1 (to be determined)
 
 * Added more props to the `Layout`
+* Allow setting and extending the user
 
 ## 0.4.0 (June 11, 2019)